requirements.txt
elasticsearch==6.8.0
fitz==0.0.1.dev2
mysql_connector_repackaged==0.3.1
pdf2image==1.16.0
Pillow==9.0.1
PyPDF2==1.26.0
pytesseract==0.3.9
python-dotenv==0.19.2
